move binary_stream to become fc::datastream
