+++ libraries/chain/include/eosio/chain/authority.hpp
@@ -20,7 +20,7 @@ struct shared_public_key {
             public_key_storage = k1r1;
          },
          [&](const shared_string& wa) {
-            fc::datastream ds(wa.data(), wa.size());
+            fc::datastream<const char*> ds(wa.data(), wa.size());
             fc::crypto::webauthn::public_key pub;
             fc::raw::unpack(ds, pub);
             public_key_storage = pub;
@@ -64,7 +64,7 @@ struct shared_public_key {
             return r1._data == r._storage.get<fc::crypto::r1::public_key_shim>()._data;
          },
          [&](const shared_string& wa) {
-            fc::datastream ds(wa.data(), wa.size());
+            fc::datastream<const char*> ds(wa.data(), wa.size());
             fc::crypto::webauthn::public_key pub;
             fc::raw::unpack(ds, pub);
             return pub == r._storage.get<fc::crypto::webauthn::public_key>();
