+++ libraries/fc/include/fc/io/bio_device_adaptor.hpp
