+++ libraries/state_history/include/eosio/state_history/compression.hpp
@@ -3,8 +3,8 @@
 #include <boost/iostreams/filtering_streambuf.hpp>
 #include <boost/iostreams/restrict.hpp>
 #include <fc/io/raw.hpp>
-#include <eosio/state_history/binary_stream.hpp>
-#include <eosio/state_history/bio_device_adaptor.hpp>
+#include <fc/io/datastream.hpp>
+#include <fc/io/bio_device_adaptor.hpp>
 
 
 namespace eosio {
@@ -48,7 +48,7 @@ void zlib_pack(STREAM& strm, const T& obj) {
    }
    else {
       length_writer<STREAM>     len_writer(strm);
-      fc::binary_stream<bio::filtering_ostreambuf> compressed_strm(bio::zlib_compressor() | fc::to_sink(strm));
+      fc::datastream<bio::filtering_ostreambuf> compressed_strm(bio::zlib_compressor() | fc::to_sink(strm));
       fc::raw::pack(compressed_strm, obj);
    }
 }
@@ -58,7 +58,7 @@ void zlib_unpack(STREAM& strm, T& obj) {
    uint32_t len;
    fc::raw::unpack(strm, len);
    if (len > 0) {
-      fc::binary_stream<bio::filtering_istreambuf> decompress_strm(bio::zlib_decompressor() | bio::restrict(fc::to_source(strm), 0, len));
+      fc::datastream<bio::filtering_istreambuf> decompress_strm(bio::zlib_decompressor() | bio::restrict(fc::to_source(strm), 0, len));
       fc::raw::unpack(decompress_strm, obj);
    }
 }
